src/main/java/org/solutions/leetcode/strings/ReverseVowelsOfAString345.java
package org.solutions.leetcode.strings;

public class ReverseVowelsOfAString345 {
    public static String reverseVowels(String s) {
        String vowels = "AEIOUaeiou";

        char[] chars = s.toCharArray();
        int left = 0, right = chars.length - 1;
        while (left < right) {
            while (left < right && vowels.indexOf(chars[left]) == -1) {
                left++;
            }
            while (left < right && vowels.indexOf(chars[right]) == -1) {
                right--;
            }
            if (left < right) {
                char temp = chars[left];
                chars[left] = chars[right];
                chars[right] = temp;
                left++;
                right--;
            }
        }
        return new String(chars);
    }
}
